refactor: import extension
